CODE/requirements.txt
anthropic>=0.36.2
jupyter>=1.1.1
google-cloud-aiplatform>=1.38
python-dotenv>=1.0.1
ollama>=0.3.3
openai>=1.52.0
pandas>=2.2.3
google-generativeai>=0.8.3
python-dotenv>=1.0.0
werkzeug>=3.0.0
click>=8.1.7
blinker>=1.7.0
scikit-learn>=1.5.2
xgboost>=2.1.2
wordcloud>=1.9.3
nltk>=3.9.1
